move method_utils into utils, since it is utils
