+++ frb_codegen/src/generator/dart/mod.rs
@@ -40,7 +40,7 @@ use convert_case::{Case, Casing};
 use log::debug;
 
 use crate::ir::IrType::*;
-use crate::method_utils::{FunctionName, MethodNamingUtil};
+use crate::utils::method::{FunctionName, MethodNamingUtil};
 use crate::others::*;
 use crate::target::Target::*;
 use crate::target::{Acc, Target};
