+++ frb_codegen/src/lib.rs
@@ -31,7 +31,6 @@ mod error;
 mod generator;
 mod ir;
 mod markers;
-mod method_utils;
 mod others;
 mod parser;
 mod target;
