+++ frb_codegen/src/parser/mod.rs
@@ -16,7 +16,7 @@ use topological_sort::TopologicalSort;
 use crate::ir::*;
 
 use crate::generator::rust::HANDLER_NAME;
-use crate::method_utils::FunctionName;
+use crate::utils::method::FunctionName;
 use crate::parser::ty::TypeParser;
 use crate::parser::source_graph::Crate;
 
@@ -334,7 +334,7 @@ fn item_method_to_function(item_impl: &ItemImpl, item_method: &ImplItemMethod) -
             Ident::new(
                 &FunctionName::new(
                     &item_method.sig.ident.to_string(),
-                    crate::method_utils::MethodInfo::Static {
+                    crate::utils::method::MethodInfo::Static {
                         struct_name: self_type.unwrap(),
                     },
                 )
@@ -345,7 +345,7 @@ fn item_method_to_function(item_impl: &ItemImpl, item_method: &ImplItemMethod) -
             Ident::new(
                 &FunctionName::new(
                     &item_method.sig.ident.to_string(),
-                    crate::method_utils::MethodInfo::NonStatic {
+                    crate::utils::method::MethodInfo::NonStatic {
                         struct_name: struct_name.clone(),
                     },
                 )
