+++ frb_codegen/src/utils/mod.rs
@@ -2,3 +2,4 @@ pub(crate) mod consts;
 pub(crate) mod misc;
 pub(crate) mod dart_repository;
 pub(crate) mod command_runner;
+pub(crate) mod method;
